requirements.txt
streamlit[pdf]>=1.45.0
qdrant-client>=1.10.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
openai>=1.30.0
requests>=2.32.0
pillow>=10.4.0
pytesseract>=0.3.13
